Add foreign key relationship between UnitMenu and User
The UnitMenu class in UnitMenu.cs now includes a new property `User` of type `int` and a new navigation property `UserNavigation` of type `User`. The User class in User.cs has been updated to include a new collection property `UnitMenus` of type `ICollection<UnitMenu>`.

The NutriAppContext class in NutriAppContext.cs has been updated to configure a new relationship between UnitMenu and User. This establishes a foreign key relationship where UnitMenu has a foreign key `User` that references the User entity. This relationship is configured to not allow deletion of the User entity if it is referenced by any UnitMenu entity, using DeleteBehavior.ClientSetNull.

Additionally, a new column named `User` of type `int` is added to the UnitMenu table, and an index is created on the User column to improve query performance. A foreign key constraint named `FK__UnitMenu__User__06CD04F7` is added, linking the User column in the UnitMenu table to the Id column in the User table.

The Down method in the migration script removes the foreign key, index, and column if the migration is rolled back. The NutriAppContextModelSnapshot file is updated to reflect these changes, including the new User column, the index on the User column, and the foreign key relationship.

Navigation properties are added to the UnitMenu and User entities to facilitate the relationship in the Entity Framework Core model.