Add many-to-many relationship between Food and MeasureType
Introduced a new entity "MeasuredBy" to establish a many-to-many relationship between "Food" and "MeasureType" entities. Updated the `NutriAppContext` to remove hardcoded seed data and use the `SeedData` class instead. Added new virtual collection properties `MeasureTypes` in `Food` and `Foods` in `MeasureType`. Created new indexes for `PracticalMeasureType` and `User` in the `FoodConsumed` entity. Updated navigation properties and added a new index on the `WaterMeasure` table for the `User` column.